不要フラグメントの削除とファイル名変更
