+++ profit-system/src/components/molecules/container/ContentBgTemplateContainer.tsx
@@ -9,10 +9,8 @@ export const ContentBgTemplate: FC<Props> = memo((props) => {
   const { children } = props;
 
   return (
-    <>
-      <Box bg="white" mb="70" borderRadius="5px" boxShadow="base">
-        {children}
-      </Box>
-    </>
+    <Box bg="white" mb="70" borderRadius="5px" boxShadow="base">
+      {children}
+    </Box>
   );
 });
