+++ profit-system/src/components/molecules/container/SbItemsContainer.tsx
@@ -9,10 +9,8 @@ export const SbItemsContainer: FC<Props> = memo((props) => {
   const { children } = props;
 
   return (
-    <>
-      <Flex direction="column" alignItems="center" mt="10px" mb="50px">
-        {children}
-      </Flex>
-    </>
+    <Flex direction="column" alignItems="center" mt="10px" mb="50px">
+      {children}
+    </Flex>
   );
 });
