+++ profit-system/src/components/pages/ProjectsList.tsx
@@ -2,9 +2,9 @@ import { Box } from "@chakra-ui/react";
 import { memo, FC } from "react";
 import { HeadLine } from "../atoms/HeadLine";
 import { MainScreenTopContainer } from "../molecules/container/MainScreenTopContainer";
-import { DateSearchSelect } from "../atoms/box/DateSearchSelect";
+import { DateSearchSelect } from "../atoms/box/DateSearchBox";
 import { PrimaryButton } from "../atoms/button/PrimaryButton";
-import { ContentBgTemplate } from "../molecules/container/ContentBgTemplate";
+import { ContentBgTemplate } from "../molecules/container/ContentBgTemplateContainer";
 
 export const ProjectsList: FC = memo(() => {
   return (
